Attestation Pool logics moved to separate class
