+++ ethereumj-core/src/main/java/org/ethereum/sharding/processing/BeaconChainFactory.java
@@ -19,6 +19,7 @@ package org.ethereum.sharding.processing;
 
 import org.ethereum.core.Block;
 import org.ethereum.db.DbFlushManager;
+import org.ethereum.sharding.crypto.Sign;
 import org.ethereum.sharding.processing.consensus.BeaconStateTransition;
 import org.ethereum.sharding.processing.consensus.GenesisTransition;
 import org.ethereum.sharding.processing.consensus.NumberAsScore;
@@ -27,6 +28,7 @@ import org.ethereum.sharding.processing.consensus.StateTransition;
 import org.ethereum.sharding.processing.db.BeaconStore;
 import org.ethereum.sharding.processing.state.BeaconState;
 import org.ethereum.sharding.processing.state.StateRepository;
+import org.ethereum.sharding.processing.validation.AttestationsValidator;
 import org.ethereum.sharding.processing.validation.BeaconValidator;
 import org.ethereum.sharding.processing.validation.StateValidator;
 import org.ethereum.sharding.pubsub.Publisher;
@@ -49,25 +51,26 @@ public class BeaconChainFactory {
 
     public static BeaconChain create(DbFlushManager beaconDbFlusher, BeaconStore store,
                                      StateRepository repository, StateTransition<BeaconState> genesisStateTransition,
-                                     StateTransition<BeaconState> stateTransitionFunction) {
+                                     StateTransition<BeaconState> stateTransitionFunction, Sign sign) {
 
         BeaconValidator beaconValidator = new BeaconValidator(store);
         StateValidator stateValidator = new StateValidator();
+        AttestationsValidator attestationsValidator = new AttestationsValidator(store, sign);
         ScoreFunction scoreFunction = new NumberAsScore();
 
-        return new BeaconChainImpl(beaconDbFlusher, store, stateTransitionFunction,
-                repository, beaconValidator, stateValidator, scoreFunction, genesisStateTransition);
+        return new BeaconChainImpl(beaconDbFlusher, store, stateTransitionFunction, repository,
+                beaconValidator, stateValidator, attestationsValidator, scoreFunction, genesisStateTransition);
     }
 
     public static BeaconChain create(DbFlushManager beaconDbFlusher, BeaconStore store, StateRepository repository,
                                      ValidatorRepository validatorRepository, Block bestBlock,
-                                     BeaconAttester beaconAttester, Publisher publisher) {
+                                     BeaconAttester beaconAttester, Publisher publisher, Sign sign) {
 
         StateTransition<BeaconState> genesisStateTransition = new GenesisTransition(validatorRepository)
                 .withMainChainRef(bestBlock.getHash());
 
         return create(beaconDbFlusher, store, repository, genesisStateTransition,
-                stateTransition(validatorRepository, beaconAttester, publisher));
+                stateTransition(validatorRepository, beaconAttester, publisher), sign);
     }
 
     public static StateTransition<BeaconState> stateTransition(ValidatorRepository validatorRepository,
