+++ ethereumj-core/src/main/java/org/ethereum/sharding/validator/BeaconProposerImpl.java
@@ -51,17 +51,17 @@ public class BeaconProposerImpl implements BeaconProposer {
     StateRepository repository;
     ValidatorConfig config;
     BeaconStore store;
-    BeaconAttester beaconAttester;
+    AttestationPool attestationPool;
 
     public BeaconProposerImpl(Randao randao, StateRepository repository, BeaconStore store,
                               StateTransition<BeaconState> stateTransition, ValidatorConfig config,
-                              BeaconAttester beaconAttester) {
+                              AttestationPool attestationPool) {
         this.randao = randao;
         this.repository = repository;
         this.store = store;
         this.stateTransition = stateTransition;
         this.config = config;
-        this.beaconAttester = beaconAttester;
+        this.attestationPool = attestationPool;
     }
 
     byte[] randaoReveal(BeaconState state, byte[] pubKey) {
@@ -83,7 +83,7 @@ public class BeaconProposerImpl implements BeaconProposer {
     public Beacon createNewBlock(Input in, byte[] pubKey) {
         Beacon lastJustified = store.getCanonicalByNumber(in.state.getCrystallizedState().getFinality().getLastJustifiedSlot());
         Beacon block = new Beacon(in.parent.getHash(), randaoReveal(in.state, pubKey), in.mainChainRef,
-                HashUtil.EMPTY_DATA_HASH, in.slotNumber, beaconAttester.getAttestations(in.slotNumber, lastJustified));
+                HashUtil.EMPTY_DATA_HASH, in.slotNumber, attestationPool.getAttestations(in.slotNumber, lastJustified));
         BeaconState newState = stateTransition.applyBlock(block, in.state);
         block.setStateHash(newState.getHash());
         logger.info("New block created {}", block);
